+++ core/src/main/java/org/trade/core/persistent/user/UserServiceImpl.java
@@ -13,7 +13,6 @@ import org.trade.core.persistent.role.RoleRepository;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 /**
@@ -36,35 +35,29 @@ public class UserServiceImpl implements UserService {
         this.passwordEncoder = passwordEncoder;
     }
 
-    @Override
     public List<User> getUsers() {
 
         return userRepository.findAll();
     }
 
-    @Override
-    public Optional<User> getUserByUsername(String username) {
+    public User getUserByUsername(String username) {
 
-        return userRepository.findByUsername(username);
+        return userRepository.findByUsername(username).orElse(null);
     }
 
-    @Override
     public boolean hasUserWithUsername(String username) {
 
         return userRepository.existsByUsername(username);
     }
 
-    @Override
     public boolean hasUserWithEmail(String email) {
 
         return userRepository.existsByEmail(email);
     }
 
-    @Override
     public User validateAndGetUserByUsername(String username) {
 
-        return getUserByUsername(username)
-                .orElseThrow(() -> new UserNotFoundException(String.format("User with username %s not found", username)));
+        return userRepository.findByUsername(username).orElseThrow(() -> new UserNotFoundException(String.format("User with username %s not found", username)));
     }
 
     @Transactional
@@ -74,9 +67,8 @@ public class UserServiceImpl implements UserService {
 
         for (Role role : instance.getRoles()) {
 
-            Role current = roleRepository.findByName(role.getName());
-
-            roles.add(Objects.requireNonNullElse(current, role));
+            Optional<Role> current = roleRepository.findByName(role.getName());
+            current.ifPresent(roles::add);
         }
 
         instance.getRoles().clear();
@@ -87,25 +79,29 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(User user) {
 
+        if (null == user) {
+
+            return;
+        }
+
         userRepository.delete(user);
     }
 
     @Override
     public Optional<User> validUsernameAndPassword(String username, String password) {
 
-        return getUserByUsername(username)
-                .filter(user -> passwordEncoder.matches(password, user.getPassword()));
+        return userRepository.findByUsername(username).filter(user -> passwordEncoder.matches(password, user.getPassword()));
     }
 
     public User findUserByName(String name) {
 
-        return userRepository.findByName(name);
+        return userRepository.findByName(name).orElse(null);
     }
 
     public UserDetails loadUserByUsername(String name) throws UsernameNotFoundException {
 
-        User user = this.userRepository.findByName(name);
-        _log.info("User found: " + user.getName() + " " + user.getPassword() + " " + String.join(",", user.getRoleValues()));
+        User user = this.userRepository.findByName(name).orElseThrow(() -> new UserNotFoundException(String.format("User with name %s not found", name)));
+        _log.info("User found: {}, password: {}, roles: {}", user.getName(), user.getPassword(), String.join(",", user.getRoleValues()));
 
         return new org.springframework.security.core.userdetails.User(user.getName(), user.getPassword(),
                 AuthorityUtils.createAuthorityList(user.getRoleValues()));
